// frontend/src/client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type FileUploadDto = {
    file: Blob | File
}

export type AnalyzeResultItemDto = {
    id: string
    name: string
    carbs: number
    cal: number
    fat: number
    protein: number
    weight: number
}

export type AnalysisResultDto = {
    items: Array<AnalyzeResultItemDto>
}

export type AnalyzeData = {
    /**
     * Image file to analyze
     */
    formData: FileUploadDto
}

export type AnalyzeResponse = AnalysisResultDto

export type $OpenApiTs = {
    '/analyze': {
        post: {
            req: {
                /**
                 * Image file to analyze
                 */
                formData: FileUploadDto
            }
            res: {
                /**
                 * Analysis result
                 */
                200: AnalysisResultDto
            }
        }
    }
}